+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog for Macro Sets WoW Addon
 
+## Version TBD - [xx/xx/2024]
+
+- Features:
+  - TBD
+- Changes:
+  - Alphabetized macro set list for easier referencing.
+- Commands:
+  - TBD
+
 ## Version 1.1.0 - [12/18/2023]
 
 - Features: